PR remark

+++ gridcapa-core-valid-day-ahead-conservative-app/src/main/java/com/farao_community/farao/gridcapa_core_valid_day_ahead_conservative/app/util/ConservativeIvaCalculationUtils.java
@@ -90,11 +90,9 @@ public final class ConservativeIvaCalculationUtils {
 
         if (hasNoTransmissionLimit(cnec)) {
             // we adjust our virtual margin by a quantity defined in task/process parameters
-            final BigDecimal inputMargin = BigDecimal.valueOf(hasNoContingency(cnec) ?
-                                                                  preventiveIvaMargin
-                                                                  : curativeIvaMargin);
+            final int inputMargin = hasNoContingency(cnec) ? preventiveIvaMargin : curativeIvaMargin;
             // adjusted margin can be < 0 given the subtraction
-            virtualMargin = virtualMargin.subtract(inputMargin).max(ZERO);
+            virtualMargin = virtualMargin.subtract(BigDecimal.valueOf(inputMargin)).max(ZERO);
         }
 
         return virtualMargin;
